Add unit tests for `LeagueService`, refactor `League` status handling to use `LeagueStatus` enum, and update repository queries accordingly.

+++ services/league/src/main/java/com/sportstock/league/repo/LeagueMemberRepository.java
@@ -16,5 +16,5 @@ public interface LeagueMemberRepository extends JpaRepository<LeagueMember, Long
 
     List<LeagueMember> findAllByLeagueId(Long leagueId);
 
-    long countByLeagueId(Long leagueId);
+    int countByLeagueId(Long leagueId);
 }
